workflows/requirements.txt
crewai==0.193.2
crewai_tools==0.76.0
openai==1.90.0
pynvml==12.0.0
easygui==0.98.3
pypandoc_binary==1.15
ruamel_yaml==0.18.14
keybert==0.9.0
hf_transfer==0.1.9
deepeval==3.6.8
evaluate==0.4.6
rouge_score==0.1.2
bert_score==0.3.13
codebleu==0.7.0
qdrant-client==1.15.1
transformers==4.56.0
peft==0.17.1
trl==0.22.2
dataframe_image==0.2.7
# pysqlite3==0.5.4
# pysqlite3-binary==0.5.4
pytesseract==0.3.13
jsonpath_ng==1.7.0
validators==0.35.0
pypdfium2==4.30.0
PyGithub==2.8.1
GitPython==3.1.46